+++ src/main.cpp
@@ -18,7 +18,6 @@
 #include "modules/plex/PlexBackend.h"
 #include "modules/jellyfin/JellyfinBackend.h"
 #include "modules/karaoke/KaraokeBackend.h"
-#include "modules/ambient_mode/AmbientModeBackend.h"
 #include "modules/tumblr_screensaver/TumblrScreensaverBackend.h"
 #include "modules/nature/NatureBackend.h"
 #include "player/MpvController.h"
@@ -134,7 +133,6 @@ int main(int argc, char *argv[]) {
     PlexBackend         plexBackend(appRoot, dataRoot);
     JellyfinBackend     jellyfinBackend(appRoot, dataRoot);
     KaraokeBackend      karaokeBackend(appRoot, dataRoot);
-    AmbientModeBackend  ambientMode(appRoot, dataRoot);
     TumblrScreensaverBackend tumblrScreensaver;
     NatureBackend       natureBackend(dataRoot);
     MpvController       mpvController(appRoot, &appCore);
@@ -171,7 +169,6 @@ int main(int argc, char *argv[]) {
     appCore.registerModule("com.240mp.plex",         "plexBackend",        &plexBackend, ctx);
     appCore.registerModule("com.240mp.jellyfin",     "jellyfinBackend",    &jellyfinBackend, ctx);
     appCore.registerModule("com.240mp.karaoke",      "karaokeBackend",     &karaokeBackend, ctx);
-    appCore.registerModule("com.240mp.ambient_mode", "ambientModeBackend", &ambientMode, ctx);
     appCore.registerModule("com.240mp.tumblr_screensaver", "tumblrScreensaverBackend", &tumblrScreensaver, ctx);
     appCore.registerModule("com.240mp.nature", "natureBackend", &natureBackend, ctx);